refactor to dockerflow controller and test it
